+++ drivers/adc/CMakeLists.txt
@@ -34,7 +34,7 @@ zephyr_library_sources_ifdef(CONFIG_ADC_GD32		adc_gd32.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_ADS1112		adc_ads1112.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_ADS1119		adc_ads1119.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_ADS7052		adc_ads7052.c)
-zephyr_library_sources_ifdef(CONFIG_ADC_ADS1X4S0X		adc_ads114s0x.c)
+zephyr_library_sources_ifdef(CONFIG_ADC_ADS1X4S0X		adc_ads1x4s0x.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_ADS131M02		adc_ads131m02.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_RPI_PICO	adc_rpi_pico.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_XMC4XXX		adc_xmc4xxx.c)
