+++ drivers/adc/Kconfig
@@ -100,7 +100,7 @@ source "drivers/adc/Kconfig.ads1119"
 
 source "drivers/adc/Kconfig.ads7052"
 
-source "drivers/adc/Kconfig.ads114s0x"
+source "drivers/adc/Kconfig.ads1x4s0x"
 
 source "drivers/adc/Kconfig.ads131m02"
