+++ drivers/adc/Kconfig.ads1x4s0x
