+++ drivers/gpio/CMakeLists.txt
@@ -7,7 +7,7 @@ zephyr_library()
 # zephyr-keep-sorted-start
 zephyr_library_sources_ifdef(CONFIG_GPIO_AD559X     gpio_ad559x.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_ADP5585    gpio_adp5585.c)
-zephyr_library_sources_ifdef(CONFIG_GPIO_ADS1X4S0X  gpio_ads114s0x.c)
+zephyr_library_sources_ifdef(CONFIG_GPIO_ADS1X4S0X  gpio_ads1x4s0x.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_ALTERA_PIO gpio_altera_pio.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_AMBIQ      gpio_ambiq.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_ANDES_ATCGPIO100 gpio_andes_atcgpio100.c)
