+++ drivers/gpio/Kconfig.ads1x4s0x
