+++ dts/bindings/gpio/ti,ads1x4s0x-gpio.yaml
