+++ include/zephyr/dt-bindings/adc/ads1x4s0x_adc.h
